% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yoga_Analysis.R
\docType{data}
\name{Yoga_Analysis}
\alias{Yoga_Analysis}
\title{A short title}
\format{
A data frame with 28 observations with 12 columns.
\describe{
\item{index}{Description of column 1, including units if appropriate}
\item{Date}{Description of column 2, including units if appropriate}
\item{Week.Day}{Description of column 2, including units if appropriate}
\item{Total.Screen.Time}{Description of column 2, including units if appropriate}
\item{Social.Networking}{Description of column 2, including units if appropriate}
\item{Reading.and.Reference}{Description of column 2, including units if appropriate}
\item{Other}{Description of column 2, including units if appropriate}
\item{Productivity}{Description of column 2, including units if appropriate}
\item{Health.and.Fitness}{Description of column 2, including units if appropriate}
\item{Entertainment}{Description of column 2, including units if appropriate}
\item{Creativity}{Description of column 2, including units if appropriate}
\item{Yoga}{Description of column 2, including units if appropriate}
}
}
\usage{
Yoga_Analysis
}
\description{
A longer paragraph that describes the context of the dataset and
discusses important aspects that will be necessary for somebody first seeing
the data to know about. Any text in these initial paragraphs will be
in the description section of the documentation file.
}
\keyword{datasets}
