% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yoga_Analysis.R
\docType{data}
\name{Yoga_Analysis}
\alias{Yoga_Analysis}
\title{Yoga and Screen Time}
\format{
A data frame with 28 observations with 12 columns.
\describe{
\item{Date}{Month, day, and year of observation.}
\item{Week.Day}{Day of week.}
\item{Total.Screen.Time}{Total amount of screen time (minutes).}
\item{Social.Networking}{Amount of time spent on social media (minutes).}
\item{Reading.and.Reference}{Amount of time spent doing reading and reference activities (minutes).}
\item{Other}{Amount of time spent on other activities (minutes).}
\item{Productivity}{Amount of time spent on productive activities, such as work (minutes).}
\item{Health.and.Fitness}{Amount of time spent working out (minutes).}
\item{Entertainment}{Amount of time spent watching movies/playing video games etc (minutes).}
\item{Creativity}{Amount of time spent creating art/music (minutes).}
\item{Yoga}{Indicates whether or not yoga was done that day.}
}
}
\source{
\url{https://www.kaggle.com/datasets/thedevastator/how-does-daily-yoga-impact-screen-time-habits}
}
\usage{
Yoga_Analysis
}
\description{
Dataset for analysis of screen time in association with other
activities including yoga, productive and creative activity.
Data is collected daily for two months.
}
\keyword{datasets}
